requirements.txt
fastapi
uvicorn
Pillow
discord-webhook
langchain
faiss-cpu
ollama
